+++ backends/windows-kernel/dualstack.go
@@ -21,12 +21,21 @@ package winkernel
 
 import (
 	"strings"
+	"time"
+	"net"
+	"fmt"
 	"k8s.io/klog/v2"
+
+	"github.com/Microsoft/hcsshim/hcn"
+	"k8s.io/client-go/tools/events"
+	"k8s.io/kubernetes/pkg/proxy/metaproxier"
+	"k8s.io/kubernetes/pkg/proxy/healthcheck"
+	"k8s.io/kubernetes/pkg/proxy"
+	"k8s.io/kubernetes/pkg/proxy/apis/config"
 
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
         kubefeatures "k8s.io/kubernetes/pkg/features"
 
-	"github.com/Microsoft/hcsshim/hcn"
 )
 
 func isOverlay(hnsNetworkInfo *hnsNetworkInfo) bool {
