+++ backends/windows-kernel/proxier_events.go
@@ -1,0 +1,91 @@
+package winkernel
+
+import (
+	v1 "k8s.io/api/core/v1"
+	discovery "k8s.io/api/discovery/v1"
+)
+
+// OnEndpointsAdd is called whenever creation of new endpoints object
+// is observed.
+func (proxier *Proxier) OnEndpointsAdd(endpoints *v1.Endpoints) {}
+
+// OnEndpointsDelete is called whenever deletion of an existing endpoints
+// object is observed.
+func (proxier *Proxier) OnEndpointsDelete(endpoints *v1.Endpoints) {}
+
+// OnServiceAdd is called whenever creation of new service object
+// is observed.
+func (proxier *Proxier) OnServiceAdd(service *v1.Service) {
+        proxier.OnServiceUpdate(nil, service)
+}
+
+// OnEndpointsUpdate is called whenever modification of an existing
+// endpoints object is observed.
+func (proxier *Proxier) OnEndpointsUpdate(oldEndpoints, endpoints *v1.Endpoints) {}
+
+
+// OnEndpointsSynced is called once all the initial event handlers were
+// called and the state is fully propagated to local cache.
+func (proxier *Proxier) OnEndpointsSynced() {}
+
+// OnServiceUpdate is called whenever modification of an existing
+// service object is observed.
+func (proxier *Proxier) OnServiceUpdate(oldService, service *v1.Service) {
+        if proxier.serviceChanges.Update(oldService, service) && proxier.isInitialized() {
+                proxier.Sync()
+        }
+}
+
+// OnServiceDelete is called whenever deletion of an existing service
+// object is observed.
+func (proxier *Proxier) OnServiceDelete(service *v1.Service) {
+        proxier.OnServiceUpdate(service, nil)
+}
+
+// OnServiceSynced is called once all the initial event handlers were
+// called and the state is fully propagated to local cache.
+func (proxier *Proxier) OnServiceSynced() {
+        proxier.mu.Lock()
+        proxier.servicesSynced = true
+        proxier.setInitialized(proxier.endpointSlicesSynced)
+        proxier.mu.Unlock()
+
+        // Sync unconditionally - this is called once per lifetime.
+        proxier.syncProxyRules()
+}
+
+// OnEndpointSliceAdd is called whenever creation of a new endpoint slice object
+// is observed.
+func (proxier *Proxier) OnEndpointSliceAdd(endpointSlice *discovery.EndpointSlice) {
+        if proxier.endpointsChanges.EndpointSliceUpdate(endpointSlice, false) && proxier.isInitialized() {
+                proxier.Sync()
+        }
+}
+
+// OnEndpointSliceUpdate is called whenever modification of an existing endpoint
+// slice object is observed.
+func (proxier *Proxier) OnEndpointSliceUpdate(_, endpointSlice *discovery.EndpointSlice) {
+        if proxier.endpointsChanges.EndpointSliceUpdate(endpointSlice, false) && proxier.isInitialized() {
+                proxier.Sync()
+        }
+}
+
+// OnEndpointSliceDelete is called whenever deletion of an existing endpoint slice
+// object is observed.
+func (proxier *Proxier) OnEndpointSliceDelete(endpointSlice *discovery.EndpointSlice) {
+        if proxier.endpointsChanges.EndpointSliceUpdate(endpointSlice, true) && proxier.isInitialized() {
+                proxier.Sync()
+        }
+}
+
+// OnEndpointSlicesSynced is called once all the initial event handlers were
+// called and the state is fully propagated to local cache.
+func (proxier *Proxier) OnEndpointSlicesSynced() {
+        proxier.mu.Lock()
+        proxier.endpointSlicesSynced = true
+        proxier.setInitialized(proxier.servicesSynced)
+        proxier.mu.Unlock()
+
+        // Sync unconditionally - this is called once per lifetime.
+        proxier.syncProxyRules()
+}
